==> 4b- load wind data.Rmd <==
---
title: "4b -load wind data"
author: "Fabian Scheifele"
date: "2023-03-10"
output: html_document
editor_options: 
  chunk_output_type: console
---
0. load packages
```{r setup, include=FALSE}
suppressMessages(memory.limit(size = NA))

if(!require(install.load)){
  install.packages("install.load")
  library(install.load)
}


suppressMessages(install_load("tidyverse", "dplyr", "ggplot2", "readr", "readxl","knitr","here", "tidyverse" , "anytime" , "data.table", "corrplot", "naniar", "fuzzyjoin","expss", "janitor","readr", "data.table", "labelled", "countrycode","grep", "economiccomplexity","xtable","ggbump", "cowplot","ggrepel", "survival","lubridate", "tidyr", "data.table", "geosphere"))

```

1.load data
```{r}
#empty data frame to bind 575 csv files together 
  meteo_data <- data.frame()

#creating list of the files, which will be uploaded  
  file_paths <-
    list.files(here("raw", "2005-2022"),
               pattern = "\\.csv$",
               full.names = TRUE)

   
    
 #file uploading 
  for (file in file_paths) {
    # read in file
    data <-
      fread(
        file,
        header = FALSE,
        fill = TRUE,
        encoding = "Latin-1",
        dec = ",",
        sep2 = ";"
      )
    
    # data manipulation: creating columns, deleting unnnecessary rows
    
    data <-
      data %>% mutate(station_name = nth(V3, 1)) %>% 
      mutate(station_code = nth(V3, 2)) %>%
      mutate(lon = nth(V2, 4)) %>%
      mutate(lat = nth(V2, 3)) %>% 
      slice(-(1:11))
    
    data <-
      data %>% separate(
        V1,
        into = c("date", "wind_speed"),
        sep = ";",
        extra = "drop"
      )
    data <- data %>% mutate(year = substr(date, 1, 4)) %>% select(-c(3:7, 1))
    
#adding separate file data to one data frame    
    
    meteo_data <- bind_rows(meteo_data, data)
  }
  
  # deleting unnecessary columns
  meteo_data <- meteo_data %>% select(-c(2, 7, 8, 9))

  #replace all non-ASCII characters with an empty string, to avoid error in numeric conversion  

meteo_data <- mutate_all(meteo_data, ~ str_replace(.x, "[^[:ascii:]]", ""))
  

#transforming columns to numeric format
  meteo_data <-
    meteo_data %>% mutate_at(c('wind_speed', 'year', 'lon', 'lat'), as.numeric)

#transforming full zeros to NAs
meteo_data$wind_speed[meteo_data$wind_speed==0]<-NA  
  
  #grouping by year, station code and location(lon and lat), averaging wind speed in a year
  meteo_data <-
    meteo_data %>% group_by(station_code, lat, lon) %>%
    summarise(avg_wind_speed = mean(wind_speed, na.rm = TRUE),
              .groups = 'keep') %>% ungroup()

  #remove stations that had complete NAs
  
meteo_data<-meteo_data%>%filter(!is.nan(avg_wind_speed))
              
```

3. merging them with the closest municipalities
```{r}
municipalities<-fread(here("raw","municipios.csv"), sep = ',', encoding = "Latin-1")%>%distinct(codigo_ibge, latitude, longitude)

library(geosphere)

# calculate the distance matrix between weather stations and municipalities
dist_matrix <- distm(meteo_data[,c("lat", "lon")], municipalities[,c("latitude", "longitude")], fun=distGeo)

# find the index of the closest weather station to each municipality
closest_station_index <- apply(dist_matrix, 2, function(x) which.min(x))

# add the station_code column to the municipalities dataframe
municipalities$station_code <- meteo_data$station_code[closest_station_index]


library(dplyr)

# left-join the meteo_data dataframe to the municipalities dataframe
merged_data <- left_join(municipalities, meteo_data, by = "station_code")%>%
  select(1:3,station_code, avg_wind_speed)

saveRDS(merged_data,here("intermediate","wind_municipalities.RDS"))

```
